+++ src/MetaFactory.php
@@ -11,24 +11,45 @@ use Lemonade\Meta\Entity\MetaEntityInterface;
 
 final class MetaFactory implements Stringable
 {
-    /** @var MetaEntityInterface[] */
-    private array $entities;
+    /** @var array<int, array<class-string<MetaEntityInterface>, MetaEntityInterface>> */
+    private array $entities = [];
 
     public function __construct(
         protected readonly MetaData $data
     ) {
-        $this->entities = [
-            new Meta($data),
-            new Facebook($data),
-            new Twitter($data),
-            new Dc($data),
-        ];
+        $this
+            ->addEntity(new Meta($this->data), 20)
+            ->addEntity(new Dc($this->data), 10)
+            ->addEntity(new Facebook($this->data), 30)
+            ->addEntity(new Twitter($this->data), 30);
+    }
+
+    public function addEntity(MetaEntityInterface $entity, int $priority = 0): self
+    {
+        $this->entities[$priority][get_class($entity)] = $entity;
+        return $this;
+    }
+
+    public function removeEntity(string $entityClassName): self
+    {
+        foreach ($this->entities as $priority => $group) {
+            if (isset($group[$entityClassName])) {
+                unset($this->entities[$priority][$entityClassName]);
+                return $this;
+            }
+        }
+        return $this;
     }
 
     public function toHtml(): string
     {
+        ksort($this->entities);
+
         return PHP_EOL
-            . implode('', array_map(fn($entity) => $entity->render(), $this->entities))
+            . implode('', array_map(
+                fn(MetaEntityInterface $entity) => $entity->render(),
+                array_merge(...array_values($this->entities))
+            ))
             . PHP_EOL;
     }